% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ploidyverse_deps}
\alias{ploidyverse_deps}
\title{List all ploidyverse dependencies}
\usage{
ploidyverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
ploidyverse packages.}
}
\description{
List all ploidyverse dependencies
}
