% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ploidyverse_packages}
\alias{ploidyverse_packages}
\title{List all packages in the ploidyverse}
\usage{
ploidyverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include ploidyverse in the list?}
}
\description{
List all packages in the ploidyverse
}
\examples{
ploidyverse_packages()
}
