% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ploidyverse_update}
\alias{ploidyverse_update}
\title{Update ploidyverse packages}
\usage{
ploidyverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
ploidyverse packages.}
}
\description{
This will check to see if all ploidyverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
ploidyverse_update()
}
}
