% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyverse.R
\docType{package}
\name{ploidyverse-package}
\alias{ploidyverse}
\alias{ploidyverse-package}
\title{ploidyverse: Install and manage packages in the ploidyverse}
\description{
The ploidyverse is a group of R packages for the analysis of genetic data
  from diploid, polyploid, and mixed-ploidy samples using standardized input and output 
  formats. The intent of the `ploidyverse` package is to provide an easy way to install, 
  load, and manage all of these packages.
}
\author{
\strong{Maintainer}: Paul Blischak \email{paul.blischak@gmail.com}

Authors:
\itemize{
  \item Lindsay V. Clark \email{lvclark@illinois.edu} (0000-0002-3881-9252)
  \item David Gerard \email{davidgerard@gmail.com}
  \item Roeland Voorrips \email{roeland.voorrips@wur.nl}
}

}
\keyword{internal}
